% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{deck}
\alias{deck}
\title{Generate a deck of cards}
\usage{
deck(decks = 1, new = FALSE)
}
\arguments{
\item{decks}{how many decks do you want?}

\item{new}{specifies a brand new unshuffled deck or number of decks}
}

