% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw a card from a deck}
\usage{
draw(cards = 5, deck_id = "new")
}
\arguments{
\item{cards}{How many cards do you want to draw}

\item{deck_id}{which deck do you want to use (defaults to \code{new})}
}

