% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_from_pile.R
\name{draw_from_pile}
\alias{draw_from_pile}
\title{Draw a card from a deck}
\usage{
draw_from_pile(cards, deck_id = "new", pile_name = "discard")
}
\arguments{
\item{cards}{How many cards do you want to draw}

\item{deck_id}{which deck do you want to use (defaults to \code{new})}

\item{pile_name}{the name of the pile to draw from (defaults to \code{discard})}
}

