% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{pile}
\alias{pile}
\title{Place cards in a pile}
\usage{
pile(cards, deck_id = "new", pile_name = "discard")
}
\arguments{
\item{cards}{a vector of the cards to place in the pile}

\item{deck_id}{identifier for deck}

\item{pile_name}{the name of the pile to create (defaults to \code{discard})}
}

